+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [unreleased]
+### Added
+- Default filter can now also be a context or project. [#86]
+
 ## [1.0.13] - 2022-04-17
 ### Added
 - TxDx's uses folder-based storage now. [#82]